backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
python-dotenv
psycopg2-binary
